+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 **Internal**:
 
+- Implement quota system for cardinality limiter. ([#2972](https://github.com/getsentry/relay/pull/2972))
 - Proactively move on-disk spool to memory. ([#2949](https://github.com/getsentry/relay/pull/2949))
 - Default missing `Event.platform` and `Event.level` fields during light normalization. ([#2961](https://github.com/getsentry/relay/pull/2961))
 - Copy event measurements to span & normalize span measurements. ([#2953](https://github.com/getsentry/relay/pull/2953))